refactor: centralize Confirmacion model
